feat: structure work (added classes CFSCalculator and CFSEngine)

+++ python/src/cfs_simp.py
@@ -2,7 +2,7 @@
 #import utils
 
 MIN_GRANULARITY = 0.75  #granularité minimale
-L = 6   #période d'ordonnancement
+L = 6.0   #période d'ordonnancement
 PRIO_TO_WEIGHT = [88761, 71755, 56483, 46273, 36291, 29154, 23254, 18705, 14949, 11916,
                   9548, 7620, 6100, 4904, 3906,
                   3121, 2501, 1991, 1586, 1277,
@@ -10,7 +10,7 @@ PRIO_TO_WEIGHT = [88761, 71755, 56483, 46273, 36291, 29154, 23254, 18705, 14949,
                   110, 87, 70, 56, 45, 36, 29, 23, 18, 15]
 
 
-def cfs_simp(tasks: list[int | list[int]], L: int=L, min_granularity: float=MIN_GRANULARITY) -> None :
+def cfs_simp(tasks: list[int | list[int]], L: float=L, min_granularity: float=MIN_GRANULARITY) -> None :
     """
     Simplified CFS scheduler.
